feat(rbac): implement user roles and add internal APIs for service communication

+++ src/main/java/com/graduationproject/authorization_service/repository/PermissionRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PermissionRepository extends JpaRepository<Permission, Integer> {
-   
+public interface PermissionRepository extends JpaRepository<Permission, Long> {
+
     Optional<Permission> findByName(String name);
 
     boolean existsByName(String name);
